improving metadatas, cleaning, creating a client component for boardArea
